+++ sources/raycast/calculetePixel.c
@@ -1,23 +1,24 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   control.c                                          :+:      :+:    :+:   */
+/*   calculetePixel.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: wwalas- <wwallas-@student.42sp.org.br>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/12/21 11:32:20 by wwallas-          #+#    #+#             */
-/*   Updated: 2023/01/07 13:04:01 by wwalas-          ###   ########.fr       */
+/*   Created: 2023/01/07 12:47:41 by wwalas-           #+#    #+#             */
+/*   Updated: 2023/01/07 12:55:33 by wwalas-          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <cub3D.h>
 
-double	player_posX(t_data *data)
-{
-	return (data->map.p_player[P_X]);
-}
+//	calculate lowest and highest pixel to fill in current stripe
 
-double	player_posY(t_data *data)
+void	calculatePixel(t_raycast *itens, t_data *data)
 {
-	return (data->map.p_player[P_Y]);
+	if (itens->drawStart < 0)
+		itens->drawStart = 0;
+	itens->drawEnd = itens->lineHeight / 2 + 480 / 2;
+	if (itens->drawEnd >= 480)
+		itens->drawEnd = 480 - 1;
 }
